Update: readme, code cleanup
