--- hra_amap.egg-info/PKG-INFO
@@ -1,5 +1,0 @@
-Metadata-Version: 2.4
-Name: hra_amap
-Version: 1.0
-License-File: LICENSE
-Dynamic: license-file
