--- hra_amap.egg-info/SOURCES.txt
@@ -1,27 +1,0 @@
-LICENSE
-README.md
-setup.py
-hra_amap.egg-info/PKG-INFO
-hra_amap.egg-info/SOURCES.txt
-hra_amap.egg-info/dependency_links.txt
-hra_amap.egg-info/top_level.txt
-hra_amap/cli/__init__.py
-hra_amap/cli/registration_stage_1.py
-hra_amap/cli/registration_stage_2.py
-hra_amap/cli/run.py
-hra_amap/cli/vtk_to_millitome.py
-hra_amap/registration/__init__.py
-hra_amap/registration/dataclass.py
-hra_amap/registration/decorators.py
-hra_amap/registration/organ.py
-hra_amap/registration/pipeline.py
-hra_amap/registration/rui.py
-hra_amap/registration/steps.py
-hra_amap/registration/tissue.py
-hra_amap/utils/__init__.py
-hra_amap/utils/contants.py
-hra_amap/utils/conversions.py
-hra_amap/utils/io.py
-hra_amap/utils/metrics.py
-hra_amap/utils/preprocess.py
-hra_amap/utils/visualization.py
